requirements.txt
pyspark==2.4.5
py4j==0.10.7
pyspark-stubs==2.4.0.post8
pyyaml==5.3.1
pyspark==2.4.5